pyproject.toml
[project]
name = "tria-aibpo"
version = "1.0.0"
description = "TRIA AI-BPO Platform - WhatsApp Order Processing System"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "TRIA AI-BPO Team" }
]

dependencies = [
    # Core Kailash SDK with Nexus and DataFlow
    "kailash[nexus,dataflow]",

    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",

    # Data Processing
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",

    # Database
    "psycopg2-binary>=2.9.9",

    # Environment Configuration
    "python-dotenv>=1.0.0",

    # OpenAI for LLM agents
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
